refactor [components] : 컴포넌트명 수정 및 컴포넌트 분리
- ConfirmButtonGroup -> ConfirmAndCancelButtonGroup 이름 변경
- ConfirmAndCancelButtonGroup 컴포넌트 파일 분리
- 합성 컴포넌트의 하위 컴포넌트명을 파스칼 케이스로 변경
- props의 옵션값 변경 및 기본 옵션 값 설정
