+++ components/src/lib/components/Modal/ConfirmModal.tsx
@@ -1,43 +1,18 @@
-import { MouseEvent } from 'react';
-
-import { BASIC_BUTTON_STYLE } from '../../constants/modal';
-import { ConfirmButtonGroupProps, ConfirmModalProps, ExtraClickAction } from '../../types/modal';
+import { ConfirmModalProps } from '../../types/modal';
 import ButtonContainer from '../ButtonContainer';
+import ConfirmAndCancelButtonGroup from '../ConfirmAndCancelButtonGroup';
 
 import CenterModal from './CenterModal';
-
-function ConfirmButtonGroup({
-  confirmButton,
-  cancelButton,
-  isConfirmButtonFirst,
-  closeModal,
-}: ConfirmButtonGroupProps) {
-  const buttonOrder = isConfirmButtonFirst ? [confirmButton, cancelButton] : [cancelButton, confirmButton];
-
-  const handleButtonClick = (e: MouseEvent<HTMLButtonElement>, extraClickAction?: ExtraClickAction) => {
-    if (extraClickAction) extraClickAction();
-    closeModal();
-  };
-  return (
-    <>
-      {buttonOrder.map(({ contents, style, extraClickAction }) => (
-        <button style={style || BASIC_BUTTON_STYLE} onClick={(e) => handleButtonClick(e, extraClickAction)}>
-          {contents}
-        </button>
-      ))}
-    </>
-  );
-}
 
 export default function ConfirmModal(props: ConfirmModalProps) {
   const {
     setOpenModal,
     title,
     contents,
-    buttonContainerJustifyContent,
+    buttonContainerJustifyContent = 'center',
     confirmButton,
     cancelButton,
-    isConfirmButtonFirst = true,
+    isConfirmButtonFirst,
   } = props;
 
   const closeModal = () => setOpenModal(false);
@@ -47,7 +22,7 @@ export default function ConfirmModal(props: ConfirmModalProps) {
       {title}
       {contents}
       <ButtonContainer $buttonContainerJustifyContent={buttonContainerJustifyContent}>
-        <ConfirmButtonGroup
+        <ConfirmAndCancelButtonGroup
           confirmButton={confirmButton}
           cancelButton={cancelButton}
           isConfirmButtonFirst={isConfirmButtonFirst}
